// api/platform-admin/client/models/CreateSubscriptionPlanDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateSubscriptionPlanDto = {
  name: string;
  description: string;
  monthlyPrice: number;
  yearlyPrice: number;
  platformTransactionFee: number;
  subscriptionFeatureIds: Array<number>;
};